+++ src/components/Modals/AlgorithmsModal/GenerationTab.tsx
@@ -1,4 +1,4 @@
-import { ParsedErrorToast, parseError } from '@/config/parsedError';
+import { ParsedErrorToasts, parseError } from '@/config/parsedError';
 import {
     DefaultBipartiteGenerationParams,
     DefaultCircleGenerationParams,
@@ -146,7 +146,7 @@ export const GenerationTab = forwardRef<GenerationTabRef>((_, ref) => {
 
     const handleRun = () => {
         if (!graph.current) {
-            addToast(ParsedErrorToast.GraphNotFound);
+            addToast(ParsedErrorToasts.GraphNotFound);
             return;
         }